Delete signupViewController on storyboard. Re-built whole signup screen presentation in order to present the signup screen if the user does not exist.
