+++ JamSpace/AppDelegate.swift
@@ -73,22 +73,22 @@ class AppDelegate: UIResponder, UIApplicationDelegate, CLLocationManagerDelegate
     Parse.setApplicationId(kApplicationID,
       clientKey: kClientKey)
     
-        let jamUser = PFUser.currentUser()?.username
-      PracticeSpace.registerSubclass()
+    let jamUser = PFUser.currentUser()?.username
+    PracticeSpace.registerSubclass()
     
-        if (jamUser != nil) {
-          let storyboard = UIStoryboard(name: "Main", bundle: nil)
-    
-//          let initialView = storyboard.instantiateViewControllerWithIdentifier("initialView") as! UITabBarController
-					
-					 let initialView = storyboard.instantiateViewControllerWithIdentifier("spacePostView") as! SpacePostViewController
-    
-          self.window?.rootViewController = initialView
-    
-        } else {
-    
-    
-    }
+//    if (jamUser != nil) {
+      let storyboard = UIStoryboard(name: "Main", bundle: nil)
+      
+      //          let initialView = storyboard.instantiateViewControllerWithIdentifier("initialView") as! UITabBarController
+      
+					 let initialView = storyboard.instantiateViewControllerWithIdentifier("initialView") as! UITabBarController
+      
+      self.window?.rootViewController = initialView
+      
+//    } else {
+//      
+//      
+//    }
     
 //    locationManager.delegate = self                
 //    locationManager.requestAlwaysAuthorization()
